Change URI structure to put paths in a param.
